Placing image button on NotesPageViewController
