+++ MacroPepelelipa/MacroPepelelipa/Controller/Notebook/Notes/ImageButtonObserver.swift
@@ -1,0 +1,14 @@
+//
+//  ImageButtonObserver.swift
+//  MacroPepelelipa
+//
+//  Created by Leonardo Oliveira on 16/10/20.
+//  Copyright © 2020 Pedro Giuliano Farina. All rights reserved.
+//
+
+import Foundation
+
+internal protocol ImageButtonObserver: class {
+    func showImageButton()
+    func hideImageButton()
+}
